+++ js/controllers/autoComplete.js
@@ -7,12 +7,10 @@
 																 |_|                   
 /////////////// ------------ USER AUTOCOMPLETE CONTROLLER ------------ ///////////////*/
 
-postworld.directive( 'userAutocomplete', [ '$log', '_', function( $log, $_ ){
+postworld.directive( 'pwUserAutocomplete', [ '$log', '_', function( $log, $_ ){
 	return {
 		controller: 'UserAutocompleteController',
-		link: function( $scope, element, attrs ){
-
-		}
+		//link: function( $scope, element, attrs ){}
 	}
 }]);
 
@@ -81,8 +79,6 @@ postworld.controller( 'UserAutocompleteController',
 }]);
 
 
-
-
 /*
   _____                    _         _                                  _      _       
  |_   _|_ _  __ _ ___     / \  _   _| |_ ___   ___ ___  _ __ ___  _ __ | | ___| |_ ___ 
@@ -91,27 +87,32 @@ postworld.controller( 'UserAutocompleteController',
    |_|\__,_|\__, |___/ /_/   \_\__,_|\__\___/ \___\___/|_| |_| |_| .__/|_|\___|\__\___|
 			|___/                                                |_|                   
 ////////// ------------ TAGS AUTOCOMPLETE CONTROLLER ------------ //////////*/
+
+postworld.directive('pwAutocompleteTags', [function () {
+	return {
+		restrict: 'A',
+		controller:'tagsAutocomplete',
+		//link: function ($scope, element, attrs) {}
+	};
+}])
+
 postworld.controller('tagsAutocomplete',
 	[ '$scope', '$filter', 'pwData',
 	function( $scope, $filter, pwData ){
 
+	$scope.tagsAutocompleteLoading = false;
+	$scope.tagsInput = []; // Object
 
-	$scope.tags_input = [];     // Array
-	$scope.tags_input_obj = []; // Object
-
-	$scope.queryTags = function () {
-		var queryTag = $scope.queryTag;
-
+	$scope.queryTags = function( viewValue ) {
 		var args = {
-			search: $scope.queryTag,
+			search: viewValue,
 			taxonomy:"post_tag"
 		}
-
-		pwData.tags_autocomplete( args ).then(
+		return pwData.tags_autocomplete( args ).then(
 			// Success
 			function(response) {
 				console.log(response.data);    
-				$scope.tagOptions = response.data;
+				return response.data;
 			},
 			// Failure
 			function(response) {
@@ -120,56 +121,46 @@ postworld.controller('tagsAutocomplete',
 		);
 	};
 
-	$scope.addTag = function(){
+	$scope.addTag = function( item ){
 		// Cycle through the tagOptions Object
-		angular.forEach( $scope.tagOptions, function( tag ){
-			if( tag.slug == $scope.queryTag ){
-				$scope.tags_input_obj.push(tag);
-				$scope.queryTag = "";
-			}           
-		});
+		$scope.tagsInput.push( item );
+		//$scope.queryTag = "";
 	}
 
 	$scope.removeTag = function( tag ){
 		// Define tags object
-		var tags = $scope.tags_input_obj;
+		var tags = $scope.tagsInput;
 
 		// Find the original tag object with the same matching slug
 		var removeTag = _.findWhere( tags, { slug: tag.slug } );
 
 		// Remove it from the tags object
-		$scope.tags_input_obj = _.without( tags, removeTag );
+		$scope.tagsInput = _.without( tags, removeTag );
 
 	}
 
-	$scope.newTag = function(){
+	$scope.newTag = function( tag ){
 		var newTag = {
-			name: $scope.queryTag,
-			slug: $scope.queryTag,
+			name: tag,
+			slug: tag,
 			};
-
-		$scope.tags_input_obj.push(newTag);
-		$scope.queryTag = "";
-		
+		$scope.tagsInput.push(newTag);
 	}
 
 	// Watch on the object with input tags
-	$scope.$watch( "tags_input_obj",
+	$scope.$watch( "tagsInput",
 		function (){
-			// When it changes, modify the tags_input object
-			$scope.tags_input = [];
-			angular.forEach( $scope.tags_input_obj, function( tag ){
-				$scope.tags_input.push(tag.slug);
+			// When it changes, modify the tagsInput object
+			var tagSlugs = [];
+			angular.forEach( $scope.tagsInput, function( tag ){
+				tagSlugs.push(tag.slug);
 			});
-
 			// Emit it's value to the parent controller
-			$scope.$emit('updateTagsInput', $scope.tags_input);
+			$scope.$emit('updateTagsInput', tagSlugs);
 		}, 1 );
 	
 	// Catch broadcast of load in tags
-	$scope.$on('postTagsObject', function(event, data) { $scope.tags_input_obj = data; });
-
-
+	$scope.$on('postTagsObject', function(event, data) { $scope.tagsInput = data; });
 
 }]);